added protection, tokens for users updating/deleting goals

+++ backend/controllers/goalController.js
@@ -1,12 +1,12 @@
 const asyncHandler = require('express-async-handler');
-
+const User = require('../models/userModel');
 const Goal = require('../models/goalModel');
 
 // @desc: get goals
 //@route: GET /api/goals
 //access: private
 const getGoals = asyncHandler(async (req, res) => {
-  const goals = await Goal.find({})
+  const goals = await Goal.find({ user: req.user.id})
 
   res.status(200).json(goals)
 })
@@ -20,7 +20,8 @@ const setGoal = asyncHandler(async (req, res) => {
     throw new Error('Please enter text')
   }
   const goal = await Goal.create({
-    text: req.body.text
+    text: req.body.text,
+    user: req.user.id,
   })
 
   res.status(200).json(goal)
@@ -37,6 +38,21 @@ const updateGoal = asyncHandler(async (req, res) => {
       res.status(400)
       throw new Error('Goal Not Found')
   }
+
+  const user = await User.findById(req.user.id)
+
+  //check for user
+  if(!user) {
+    res.status(401)
+    throw new Error('User not found')
+  }
+
+  //make sure the logged in user matches the user that created the goal
+  if(goal.user.toString() !== user.id) {
+    res.status(401)
+    throw new Error('User not authorized')
+  }
+
   // finding goal and updating it
   const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {new: true} )
   res.status(200).json(updatedGoal)
@@ -52,6 +68,21 @@ const deleteGoal = asyncHandler(async (req, res) => {
       res.status(400)
       throw new Error('Goal Not Found')
   }
+
+  const user = await User.findById(req.user.id)
+
+  //check for user
+  if(!user) {
+    res.status(401)
+    throw new Error('User not found')
+  }
+
+  //make sure the logged in user matches the user that created the goal
+  if(goal.user.toString() !== user.id) {
+    res.status(401)
+    throw new Error('User not authorized')
+  }
+
 
   await goal.deleteOne();
